# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PGP Key Manager API
  version: 1.0.0
  contact:
    name: PGP Key Manager
    url: https://github.com/bbruneel/PGP-key-manager
  license:
    name: Unlicense
    url: https://opensource.org/licenses/Unlicense
  description: |
    Versioned JSON API for OpenPGP key management.

    **Versioning:** `Accept: application/json; version=1`
    **Auth:** Bearer JWT (Auth0) on `/api/keys/**`, `/api/groups/**`, `/api/invites/**`, and `/api/admin/**`
    **Crypto:** Bouncy Castle (`bcprov-jdk18on`, `bcpg-jdk18on` 1.84). Primary generation supports `openpgpVersion` 4 (default) or 6.
    **Errors:** RFC 7807 `application/problem+json`
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/keys:
    get:
      tags: [Keys]
      summary: List keys
      operationId: listKeys
      x-implementation-status: implemented
      parameters:
        - name: groupId
          in: query
          schema:
            type: string
            format: uuid
        - name: scope
          in: query
          schema:
            type: string
            enum: [all, personal, group]
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/KeyRole'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/KeyStatus'
        - name: capability
          in: query
          schema:
            $ref: '#/components/schemas/PgpCapability'
      responses:
        '200':
          description: Key list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PgpKeySummary'
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags: [Keys]
      summary: Create key
      operationId: createKey
      x-implementation-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePgpKeyRequest'
      responses:
        '201':
          description: Primary key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PgpKey'
        '200':
          description: |
            Existing primary key with the same fingerprint was found. Stored armored material was
            updated and keyring metadata synced (primary/subkey revocation, missing subkey rows).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PgpKey'
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
  /api/keys/preview:
    post:
      tags: [Keys]
      summary: Preview keyring import
      operationId: previewKeyring
      x-implementation-status: implemented
      description: |
        Parse armored key material without persisting. Returns primary and subkey metadata,
        revocation state read from the keyring, and warnings when public/private blocks differ.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePgpKeyRequest'
      responses:
        '200':
          description: Parsed keyring preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreviewKeyringResponse'
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
  /api/keys/{keyId}:
    get:
      tags: [Keys]
      summary: Get key
      operationId: getKey
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/KeyId'
      responses:
        '200':
          description: Key details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PgpKey'
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
    patch:
      tags: [Keys]
      summary: Update key
      operationId: updateKey
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/KeyId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePgpKeyRequest'
      responses:
        '200':
          description: Updated key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PgpKey'
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags: [Keys]
      summary: Delete key
      operationId: deleteKey
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/KeyId'
      responses:
        '204':
          description: Deleted
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
  /api/keys/{primaryKeyId}/subkeys:
    get:
      tags: [Subkeys]
      summary: List subkeys
      operationId: listSubkeys
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/PrimaryKeyId'
      responses:
        '200':
          description: Subkey list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PgpKeySummary'
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags: [Subkeys]
      summary: Create subkey
      operationId: createSubkey
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/PrimaryKeyId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubkeyRequest'
      responses:
        '201':
          description: Subkey created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PgpKey'
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
  /api/keys/{primaryKeyId}/subkeys/import-from-keyring/preview:
    post:
      tags: [Subkeys]
      summary: Preview import subkeys from stored keyring
      operationId: previewImportSubkeysFromKeyring
      x-implementation-status: implemented
      description: |
        Parse the primary's stored armored keyring and show which subkey rows would be
        registered, updated (revocation sync), or skipped without writing to the database.
        Active primary rows whose stored keyring shows revocation appear in `wouldUpdate`
        with `role: primary`.
      parameters:
        - $ref: '#/components/parameters/PrimaryKeyId'
      responses:
        '200':
          description: Import preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreviewImportSubkeysResponse'
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
  /api/keys/{primaryKeyId}/subkeys/import-from-keyring:
    post:
      tags: [Subkeys]
      summary: Import subkey rows from stored keyring
      operationId: importSubkeysFromKeyring
      x-implementation-status: implemented
      description: |
        Parses non-master keys from the primary's stored armored keyring and registers metadata-only
        subkey rows for any fingerprints not already registered under this primary. Idempotent —
        existing subkey rows that are already up to date are skipped. Active primary and subkey rows
        whose fingerprints are revoked in the stored keyring are synced to revoked (`updated` /
        `updatedCount`, including `role: primary` for the primary row). Returns `201` when new rows
        are registered, `200` when only sync or skip occurred.
      parameters:
        - $ref: '#/components/parameters/PrimaryKeyId'
      responses:
        '200':
          description: |
            Idempotent sync completed with no new subkey rows (`registered` is empty). All keyring
            subkeys were already registered, or the keyring has no subkeys.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportSubkeysResponse'
        '201':
          description: One or more new subkey rows were registered from the keyring.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportSubkeysResponse'
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
  /api/keys/{primaryKeyId}/subkeys/{subkeyId}:
    get:
      tags: [Subkeys]
      summary: Get subkey
      operationId: getSubkey
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/PrimaryKeyId'
        - $ref: '#/components/parameters/SubkeyId'
      responses:
        '200':
          description: Subkey details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PgpKey'
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
  /api/keys/{keyId}/revoke:
    post:
      tags: [Key lifecycle]
      summary: Revoke key
      operationId: revokeKey
      x-implementation-status: implemented
      description: |
        Revokes a primary or subkey. When the account's primary key row stores encrypted private
        material, `passphrase` is required and the OpenPGP keyring is updated cryptographically.
        Without stored private material, only database metadata (`revokedAt`, `revocationReason`) is
        updated. Supplying a passphrase when no private material exists returns 400.
      parameters:
        - $ref: '#/components/parameters/KeyId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeKeyRequest'
      responses:
        '200':
          description: Revoked key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PgpKey'
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
  /api/keys/{keyId}/extend-expiry:
    post:
      tags: [Key lifecycle]
      summary: Extend key expiry
      operationId: extendKeyExpiry
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/KeyId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtendExpiryRequest'
      responses:
        '200':
          description: Key with updated expiry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PgpKey'
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
  /api/keys/{keyId}/rotate:
    post:
      tags: [Key lifecycle]
      summary: Rotate subkey
      operationId: rotateKey
      x-implementation-status: implemented
      description: |
        Creates a replacement subkey on the parent primary. `passphrase` is required (unlocks the
        primary keyring for new subkey creation; rotate is only supported when the primary stores
        private material). When `revokePrevious` is true (default), the previous subkey is also
        revoked in the OpenPGP keyring; when false, the previous subkey row is left unchanged in
        the keyring.
      parameters:
        - $ref: '#/components/parameters/KeyId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RotateKeyRequest'
      responses:
        '201':
          description: Rotation result with new and previous subkeys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RotateKeyResponse'
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
  /api/keys/{keyId}/transfer-ownership:
    post:
      tags: [Key lifecycle]
      summary: Transfer key ownership
      operationId: transferOwnership
      x-implementation-status: implemented
      description: |
        Transfers ownership between personal and team vault scopes. When `ownerGroupId` is provided,
        the key becomes group-owned. When omitted, ownership returns to the current user.
        Authorization parity with other key operations: any current key operator may transfer ownership,
        and transfer to a group requires membership in the target group.
      parameters:
        - $ref: '#/components/parameters/KeyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferOwnershipRequest'
      responses:
        '200':
          description: Updated key ownership
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PgpKey'
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
  /api/keys/{keyId}/export-public:
    get:
      tags: [Keys]
      summary: Export public key
      operationId: exportPublicKey
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/KeyId'
      responses:
        '200':
          description: Armored public key material
          content:
            application/pgp-keys:
              schema:
                type: string
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
  /api/keys/{keyId}/export-ssh-public:
    get:
      tags: [Keys]
      summary: Export OpenSSH public key
      description: |
        Returns a one-line OpenSSH public key for keys with the `authenticate` capability
        and an SSH-compatible algorithm (`ed25519`, `rsa`, or `ecdsa`). Encryption-only
        algorithms (for example `cv25519`) and keys without `authenticate` return 400.
      operationId: exportSshPublicKey
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/KeyId'
      responses:
        '200':
          description: OpenSSH authorized_keys line (type, base64, comment)
          content:
            text/plain:
              schema:
                type: string
                example: ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIExampleKeyComment openpgp:0x70dfa6b7
        '4XX':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ErrorResponse:
      description: RFC 7807 problem details (typical 400, 401, 404, or 409)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  parameters:
    KeyId:
      name: keyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PrimaryKeyId:
      name: primaryKeyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SubkeyId:
      name: subkeyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    KeyRole:
      type: string
      enum: [primary, subkey]
    KeyStatus:
      type: string
      enum: [active, expired, revoked]
    PgpCapability:
      type: string
      enum: [certify, sign, encrypt, authenticate]
    MaterialKind:
      type: string
      enum: [public, private]
...
    CreatePgpKeyRequest:
      type: object
      description: |
        Dual-purpose body for `POST /api/keys`.

        **Generate path:** include `algorithmSpec`, `passphrase`, and optional `userIds` / `validity`.
        **Register/import path:** include `armoredPublic` and/or `encryptedPrivateArmored`; the server parses
        metadata (fingerprint, key ID, algorithm, capabilities, expiry) from the armored material.
        When the armored keyring contains subkeys, metadata-only subkey rows are registered automatically
        under the new primary (same storage model as create-subkey — armored material stays on the primary row).
        `fingerprint` is optional on register — when omitted the server derives it; when provided it must match.
        When both armored blocks are sent they must refer to the same key or the request is rejected.
        Client-supplied `capabilities`, `algorithm`, `keyId`, and `expiresAt` are ignored on register;
        parsed values from the armored material are always stored.
      properties:
        label:
          type: string
        fingerprint:
          type: string
          pattern: '^[0-9A-Fa-f]{16,40}$'
          description: |
            Optional on register/import. When omitted, derived from armored key material.
            When provided, must match the parsed fingerprint or the request is rejected.
        keyType:
          $ref: '#/components/schemas/MaterialKind'
        algorithmSpec:
          $ref: '#/components/schemas/AlgorithmSpec'
        capabilities:
          type: array
          items:
            $ref: '#/components/schemas/PgpCapability'
        validity:
          $ref: '#/components/schemas/ValiditySpec'
        userIds:
          type: array
          items:
            $ref: '#/components/schemas/UserIdSpec'
        passphrase:
          $ref: '#/components/schemas/PassphraseField'
        openpgpVersion:
          type: integer
          enum: [4, 6]
          default: 4
          description: |
            OpenPGP key packet version for generated primary keys (RFC 4880 v4 or RFC 9580 v6).
            Omitted defaults to 4. Subkeys inherit the primary key version.
        armoredPublic:
          type: string
          description: Required for register/import unless `encryptedPrivateArmored` is provided.
        encryptedPrivateArmored:
          type: string
          description: Required for private register/import when `armoredPublic` is omitted.
        ownerGroupId:
          type: string
          format: uuid
          description: Optional group owner for team vault keys. Omit for personal ownership.
    CreateSubkeyRequest:
      type: object
      required: [capabilities, algorithm]
      properties:
        capabilities:
          type: array
          items:
            $ref: '#/components/schemas/PgpCapability'
        algorithm:
          $ref: '#/components/schemas/AlgorithmSpec'
        validity:
          $ref: '#/components/schemas/ValiditySpec'
        passphrase:
          $ref: '#/components/schemas/PassphraseField'
    UpdatePgpKeyRequest:
      type: object
      properties:
        label:
          type: string
        expiresAt:
          type: string
          format: date-time
        storageProvider:
          type: string
        storageRef:
          type: string
...
    RevokeKeyRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          enum: [no_reason, key_superseded, key_compromised, key_retired, user_id_invalid]
        description:
          type: string
        passphrase:
          $ref: '#/components/schemas/PassphraseField'
    ExtendExpiryRequest:
      type: object
      required: [expiresAt]
      properties:
        expiresAt:
          type: string
          format: date-time
        passphrase:
          $ref: '#/components/schemas/PassphraseField'
    RotateKeyRequest:
      type: object
      required: [capabilities, algorithm]
      properties:
        capabilities:
          type: array
          items:
            $ref: '#/components/schemas/PgpCapability'
        algorithm:
          $ref: '#/components/schemas/AlgorithmSpec'
        validity:
          $ref: '#/components/schemas/ValiditySpec'
        passphrase:
          allOf:
            - $ref: '#/components/schemas/PassphraseField'
          description: |
            Required for rotate. Unlocks the primary keyring for new subkey creation (and for
            revoking the previous subkey when `revokePrevious` is true).
        revokePrevious:
          type: boolean
          description: |
            When true (default), revokes the previous subkey in the OpenPGP keyring before creating
            the replacement. Does not affect whether `passphrase` is required.
    RotateKeyResponse:
      type: object
      required: [newKey, previousKey]
      properties:
        newKey:
          $ref: '#/components/schemas/PgpKey'
        previousKey:
          $ref: '#/components/schemas/PgpKey'
    ImportSubkeysResponse:
      type: object
      required: [registered, skippedCount, updated, updatedCount]
      properties:
        registered:
          type: array
          items:
            $ref: '#/components/schemas/PgpKeySummary'
          description: Newly registered subkey rows from this import.
        skippedCount:
          type: integer
          description: Subkeys already registered under this primary that were skipped.
        updated:
          type: array
          items:
            $ref: '#/components/schemas/PgpKeySummary'
          description: Existing primary or subkey rows synced to revoked status from the keyring.
        updatedCount:
          type: integer
          description: Number of existing rows updated (revocation sync, including primary).
...
    PreviewKeyringResponse:
      type: object
      required: [primary, subkeys, warnings, source]
      properties:
        primary:
          $ref: '#/components/schemas/PreviewKeyEntry'
        subkeys:
          type: array
          items:
            $ref: '#/components/schemas/PreviewKeyEntry'
        warnings:
          type: array
          items:
            type: string
        source:
          type: string
          enum: [public, private, both]
          description: Which armored block(s) were used; when both are pasted, private is authoritative.
    PreviewImportSubkeysResponse:
      type: object
      required: [wouldRegister, wouldUpdate, wouldSkipCount, warnings, source]
      properties:
        wouldRegister:
          type: array
          items:
            $ref: '#/components/schemas/PreviewKeyEntry'
        wouldUpdate:
          type: array
          items:
            $ref: '#/components/schemas/PreviewKeyEntry'
        wouldSkipCount:
          type: integer
        warnings:
          type: array
          items:
            type: string
        source:
          type: string
          enum: [public, private, both]
...
    PgpKeySummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        label:
          type: [string, 'null']
          description: User-facing key name. `null` when unset.
        fingerprint:
          type: string
        keyId:
          type: string
          pattern: '^[0-9A-F]{16}$'
          description: OpenPGP key ID (16 uppercase hex characters).
        algorithm:
          type: string
          description: |
            Stored algorithm identifier (e.g. `ed25519`, `cv25519`). Parsed from armored material on
            register or set on generation.
        keyType:
          $ref: '#/components/schemas/MaterialKind'
        role:
          $ref: '#/components/schemas/KeyRole'
        parentKeyId:
          type: [string, 'null']
          format: uuid
        ownerType:
          type: string
          enum: [user, group]
        ownerGroupId:
          type: [string, 'null']
          format: uuid
        capabilities:
          type: array
          items:
            $ref: '#/components/schemas/PgpCapability'
          description: Always present in responses; empty array when no capabilities are stored.
        status:
          $ref: '#/components/schemas/KeyStatus'
        expiresAt:
          type: [string, 'null']
          format: date-time
          description: |
            Key expiration instant. `null` means the key does not expire (no OpenPGP key-expiration subpacket).
        revokedAt:
          type: [string, 'null']
          format: date-time
        openpgpVersion:
          type: integer
          enum: [4, 6]
          description: OpenPGP packet version for this keyring (primary sets it; subkeys inherit).
        hasPrivateMaterial:
          type: boolean
          description: True when this key row has private key material available for cryptographic actions.
    PgpKey:
      allOf:
        - $ref: '#/components/schemas/PgpKeySummary'
        - type: object
          description: |
            Primary keys store the full armored keyring. Subkey rows store metadata only; use the
            parent primary or export-public for armored public material.
          properties:
            armoredPublic:
              type: string
              description: Present on primary keys; omitted on subkey list/detail responses.
            encryptedPrivateArmored:
              type: string
              description: Present on primary detail when generated or imported with private material.
            registeredSubkeyCount:
              type: integer
              description: |
                Present on primary register responses when subkey rows were auto-imported from a
                multi-key armored export. Omitted on generate and on other responses.